Setup  index.html and one recipe page
Set up the html boilerplate for both index.html and korean-ground-beef.html, linked the recipe page to the index page, and filled content out for recipe page.

So far had meaningful progress in code to save to codebase. I have completed content for the recipe page that will be repeated in the coming changes.